Split the docker code from the Juju plugin-specific code.
